+++ lib/src/waveform/mod.rs
@@ -1,3 +1,4 @@
 pub mod harsh;
 pub mod perfect;
 pub mod soft;
+pub mod vocal;